model/requirements.txt
torch>=2.0.0
torchvision
ftfy
tensorboard
Jinja2
peft==0.7.0
diffusers>=0.16.0
transformers>=4.25.1
accelerate>=0.16.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
pillow>=9.5.0
python-multipart>=0.0.6
